+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - performance improvements related to rendered messages
 - omit showing empty Thinking block when there is no thinking content
+- loading current question in browser client and added event when answering question
 
 ## [0.37.1]